Error if program size exceeded
* Throw exception if a command is added to userland or a command
implementation when no more commands are allowed in that program.
* Maximum program is either given by technical constraints (maximum
positive 32-bit integer value) or by program counter length (whichever
is smaller).
** In case of command implementations without explicit program counter
length, the technical limit (as above) is always used.

* Add ConstraintException
** I've decided to try to avoid throwing ParseError outside the Parser.
ConstraintException is used as a stand-in such that first
ConstraintException is thrown, then the Parser catches and replaces it
with a ParseError exception.