Ensure Variables exist when being used
* Before: Invocation.resolve() couldn't distinguish where in an
implementation a Local Variable is defined, thus an invocation may refer
to a variable that doesn't exist yet.
* Now: Only Registers, Parameters, and Local Variables that exist at the
location of invocation are visible therein.
** This is done indirectly by tentatively resolving a potential Variable
reference upon initial parsing of the invocation, instead of when the
invocation is resolved.
** Of course, it may still turn out to be a Label instead;
Invocation.resolve() has full control over figuring this out (or will,
once Label support is implemented).

* Invocation: Simplify
* RawArgument: Store potentialRegister

* Remove Argument.fromToken() (was only used in one place and would
require further modifications tailor-made for that usage)
** Moved code to Parser