Invocations have Origin
* In order to improve error messages, let's provide Origin info in more
cases (especially at runtime).

* Invocations contain Origin value so that RuntimeErrors can include it
* Redefine meaning of ConstraintException; use it to bubble up error
messages to the location (i.e. CommandInvocation) where Origin info can
be added
** I.e. Implementations and Built-in Interpretables may throw both
ConstraintExceptions and RuntimeError, but Invocations must only throw
RuntimeErrors

* Built-ins: Throw ConstraintExceptions - this way their errors are
effectively associated to the Invocation's Origin
** BuiltInFunction: Add getValue(), Add getNumericValue() to transform
ConstraintExceptions to RuntimeExceptions (as here, this error shouldn't
happen and is not a UserError)

* Origin: Add merge()

* Constraints: Remove some unused method variants
* Parser: Remove EOF cases that cannot be reached (because Tokenizer
emits a STATEMENT_SEPARATOR at the end of the last line)